MINUS SIGN - curves collapse but still very low z
All curves collapse well showing S universality. Still extracted value of z is low so want to do a size convergence study as discussed in previously sent out email - will be done by tomorrow for sure!
Still want to discuss implications of how we have set up our array of k as I don't fully get why it scales with N and not 1/N. But it seems to work correctly this way. Perhaps something to do with the grid of k being independant as it's just used to find indices, but then we use k array to average for <k> and there it confuses me why that used k still works.
Can we discuss this tonight please? @AdenLS @WillStevey @jpw447