Big N figures
Still want to discuss used k values. Especially as this L(t) is very much not right. It's average domain size, so can't be less than 1, should be ~100.  And still want to do a convergence study with N as high N does give values closer to expected 0.5, even if errors are small.