Last Data processed and saved figures
Need to emphasise to Mike and in our discussion that repeats over initial conditions are super important. Results vary a lot.
I checked by hand on N=64 and between 2 runs, results only stop varying to 3dps by 160 repeats or so. But we will have to run with what we have at this point - just report limitations